# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BitByBit Habits API
  description: |
    API REST de BitByBit, un habit tracker que recompensa completar tareas con sats reales via Lightning Network.

    **Caso de uso MVP**: "Mesada 2.0" donde sponsors (padres/tutores) crean habitos con recompensa en sats
    y kids los completan para ganar.

    ## Autenticacion

    La API usa autenticacion basada en cookies de sesion (`session`). Para autenticarse:

    1. Registrar un usuario con `POST /api/auth/register`
    2. Iniciar sesion con `POST /api/auth/login`
    3. La cookie `session` se setea automaticamente en la respuesta

    Todos los endpoints (excepto auth) requieren una sesion activa.

    ## Formato de respuesta

    Todas las respuestas siguen el formato `ApiResponse`:
    ```json
    {
      "success": true,
      "data": { ... }
    }
    ```
    En caso de error:
    ```json
    {
      "success": false,
      "error": "Descripcion del error"
    }
    ```
  version: 1.0.0
  contact:
    name: BitByBit Team
servers:
  - url: /api
    description: Servidor de la API
...
paths:
...
  /auth/session:
    get:
      tags:
        - Auth
      summary: Obtener sesion actual
      description: Devuelve los datos de la sesion del usuario autenticado.
      operationId: getSession
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Sesion activa encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
              example:
                success: true
                data:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "padre@ejemplo.com"
                  username: "papajuan"
                  display_name: "Juan"
                  role: "sponsor"
                  locale: "es"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
...
  /completions:
    get:
      tags:
        - Completions
      summary: Listar completaciones del usuario
      description: |
        Devuelve las completaciones del usuario autenticado.
        Se puede filtrar por rango de fechas usando los query parameters `from` y `to`.
      operationId: getCompletions
      security:
        - cookieAuth: []
      parameters:
        - name: from
          in: query
          description: Fecha de inicio del rango (formato YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-01"
        - name: to
          in: query
          description: Fecha de fin del rango (formato YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-07"
      responses:
        "200":
          description: Lista de completaciones obtenida exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Completion"
              example:
                success: true
                data:
                  - id: "990e8400-e29b-41d4-a716-446655440004"
                    habit_id: "660e8400-e29b-41d4-a716-446655440001"
                    user_id: "880e8400-e29b-41d4-a716-446655440003"
                    date: "2026-03-07"
                    status: "approved"
                    note: "Listo!"
                    completed_at: "2026-03-07T08:30:00.000Z"
                    reviewed_by: "550e8400-e29b-41d4-a716-446655440000"
                    reviewed_at: "2026-03-07T09:00:00.000Z"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "500":
          description: Error al obtener las completaciones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al obtener las completaciones"

    post:
      tags:
        - Completions
      summary: Registrar completacion de un habito
      description: |
        Registra que un kid completo un habito para el dia de hoy.
        Solo usuarios con rol `kid` pueden usar este endpoint.

        - Si el habito tiene `verification_type: "self_verify"`, la completacion se aprueba automaticamente.
        - Si tiene `verification_type: "sponsor_approval"`, queda en estado `pending`.
        - No se puede completar el mismo habito mas de una vez por dia.
      operationId: createCompletion
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionRequest"
            example:
              habit_id: "660e8400-e29b-41d4-a716-446655440001"
              note: "Listo!"
              evidence_url: "https://ejemplo.com/foto.jpg"
      responses:
        "201":
          description: Completacion registrada exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Completion"
              example:
                success: true
                data:
                  id: "990e8400-e29b-41d4-a716-446655440004"
                  habit_id: "660e8400-e29b-41d4-a716-446655440001"
                  user_id: "880e8400-e29b-41d4-a716-446655440003"
                  date: "2026-03-07"
                  status: "pending"
                  note: "Listo!"
                  evidence_url: "https://ejemplo.com/foto.jpg"
                  completed_at: "2026-03-07T08:30:00.000Z"
        "400":
          description: Falta el campo habit_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "El campo habit_id es obligatorio"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "403":
          description: Solo los kids pueden completar habitos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Solo los kids pueden completar habitos"
        "404":
          description: Habito no encontrado o no asignado al usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Habito no encontrado o no asignado a vos"
        "409":
          description: Ya se completo este habito hoy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Ya completaste este habito hoy"
        "500":
          description: Error al registrar la completacion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al registrar la completacion"
...
  /families/join:
    post:
      tags:
        - Families
      summary: Unirse a una familia con codigo de invitacion
      description: |
        Permite a un usuario unirse a una familia existente usando el codigo de invitacion.
        El usuario se agrega como miembro con su rol actual (sponsor o kid).
        No se puede unir a una familia de la que ya es miembro.
      operationId: joinFamily
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinFamilyRequest"
            example:
              invite_code: "ABC123"
      responses:
        "201":
          description: Union exitosa a la familia
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          family:
                            $ref: "#/components/schemas/Family"
                          member:
                            $ref: "#/components/schemas/FamilyMember"
              example:
                success: true
                data:
                  family:
                    id: "770e8400-e29b-41d4-a716-446655440002"
                    name: "Familia Rodriguez"
                    invite_code: "ABC123"
                    created_by: "550e8400-e29b-41d4-a716-446655440000"
                    created_at: "2026-03-01T10:00:00.000Z"
                  member:
                    id: "bbb00000-0000-0000-0000-000000000001"
                    family_id: "770e8400-e29b-41d4-a716-446655440002"
                    user_id: "880e8400-e29b-41d4-a716-446655440003"
                    role: "kid"
                    joined_at: "2026-03-07T10:00:00.000Z"
        "400":
          description: Codigo de invitacion faltante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "El codigo de invitacion es obligatorio"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "No autenticado"
        "404":
          description: Codigo de invitacion invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Codigo de invitacion invalido"
        "409":
          description: Ya es miembro de esta familia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Ya sos miembro de esta familia"
        "500":
          description: Error al unirse a la familia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              example:
                success: false
                error: "Error al unirse a la familia"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sesion httpOnly seteada por el endpoint de login
  schemas:
...
    CompletionRequest:
      type: object
      required:
        - habit_id
      properties:
        habit_id:
          type: string
          format: uuid
          description: ID del habito a completar
        note:
          type: string
          description: Nota opcional sobre la completacion
        evidence_url:
          type: string
          format: uri
          description: URL de evidencia opcional (foto, etc.)
...
    JoinFamilyRequest:
      type: object
      required:
        - invite_code
      properties:
        invite_code:
          type: string
          description: Codigo de invitacion de 6 caracteres (case-insensitive)
          example: "ABC123"

    # ---- Response / Model schemas ----
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indica si la operacion fue exitosa
        data:
          description: Datos de la respuesta (presente cuando success es true)
        error:
          type: string
          description: Mensaje de error (presente cuando success es false)
    ApiResponseError:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          description: Mensaje de error descriptivo
...
    AuthSession:
      type: object
      description: Datos de la sesion del usuario autenticado
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum:
            - sponsor
            - kid
        locale:
          type: string
          enum:
            - es
            - en
...
    Completion:
      type: object
      description: Registro de completacion de un habito por un kid
      properties:
        id:
          type: string
          format: uuid
        habit_id:
          type: string
          format: uuid
          description: ID del habito completado
        user_id:
          type: string
          format: uuid
          description: ID del kid que completo
        date:
          type: string
          format: date
          description: Fecha de la completacion (YYYY-MM-DD)
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
          description: Estado de la completacion
        evidence_url:
          type: string
          format: uri
          nullable: true
          description: URL de evidencia
        note:
          type: string
          nullable: true
          description: Nota del kid
        completed_at:
          type: string
          format: date-time
          description: Timestamp de cuando se registro la completacion
        reviewed_by:
          type: string
          format: uuid
          nullable: true
          description: ID del sponsor que reviso
        reviewed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp de la revision
...
    Family:
      type: object
      description: Grupo familiar
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Nombre de la familia
        invite_code:
          type: string
          description: Codigo de invitacion de 6 caracteres
        created_by:
          type: string
          format: uuid
          description: ID del sponsor que la creo
        created_at:
          type: string
          format: date-time
    FamilyMember:
      type: object
      description: Miembro de una familia
      properties:
        id:
          type: string
          format: uuid
        family_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - sponsor
            - kid
        joined_at:
          type: string
          format: date-time
...